Переносит информация табов в отдельные компоненты
