+++ src/components/movie-page/movie-page.jsx
@@ -3,34 +3,15 @@ import PropTypes from 'prop-types';
 import Tabs from "./../tabs/tabs.jsx";
 import MovieList from "./../movie-list/movie-list.jsx";
 import {TABS} from "./../../utils.js";
+import TabOverview from "./../tab-overview/tab-overview.jsx";
+import TabDetails from "./../tab-details/tab-details.jsx";
+import TabReviews from "./../tab-reviews/tab-reviews.jsx";
 
 const DISPLAYED_NUMBER_OF_FILMS = 4;
-
-const getRatingLevel = (rating) => {
-  const formattedRating = parseFloat(rating.replace(`,`, `.`));
-
-  if (formattedRating === 0 || formattedRating < 3) {
-    return `Bad`;
-  }
-  if (formattedRating >= 3 || formattedRating < 5) {
-    return `Normal`;
-  }
-  if (formattedRating >= 5 || formattedRating < 8) {
-    return `Good`;
-  }
-  if (formattedRating >= 8 || formattedRating < 10) {
-    return `Very good`;
-  }
-  if (formattedRating >= 10) {
-    return `Awesome`;
-  }
-  return null;
-};
 
 const MoviePage = (props) => {
   const {film, onFilmTitleClick, onFilmImgClick, activeItem, onClick} = props;
   const {title, posterSrc, movieCoverSrc, genre, yearRelease, description, rating, numberVotes, producer, actors, runTime} = film;
-  const ratingLevel = getRatingLevel(rating);
   const listTabs = Object.values(TABS);
 
   return <React.Fragment>
@@ -101,117 +82,15 @@ const MoviePage = (props) => {
             </nav>
 
             {activeItem === TABS.overview &&
-            <React.Fragment>
-              <div className="movie-rating">
-                <div className="movie-rating__score">{rating}</div>
-                <p className="movie-rating__meta">
-                  <span className="movie-rating__level">{ratingLevel}</span>
-                  <span className="movie-rating__count">{numberVotes} ratings</span>
-                </p>
-              </div>
-              <div className="movie-card__text">
-                <p>{description}</p>
-
-                <p className="movie-card__director"><strong>Director: {producer}</strong></p>
-
-                <p className="movie-card__starring"><strong>Starring: {actors.join(`, `)}</strong></p>
-              </div>
-            </React.Fragment>}
+            <TabOverview film={film}/>
+            }
 
             {activeItem === TABS.details &&
-              <div className="movie-card__text movie-card__row">
-                <div className="movie-card__text-col">
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Director</strong>
-                    <span className="movie-card__details-value">{producer}</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Starring</strong>
-                    <span className="movie-card__details-value">
-                      {actors.map((actor, i, arr)=>{
-                        return <React.Fragment
-                          key={actor}>
-                          {actor}{i === arr.length - 1 ? `` : `,`} <br />
-                        </React.Fragment>;
-                      })}
-                    </span>
-                  </p>
-                </div>
-
-                <div className="movie-card__text-col">
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Run Time</strong>
-                    <span className="movie-card__details-value">{runTime}</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Genre</strong>
-                    <span className="movie-card__details-value">{genre}</span>
-                  </p>
-                  <p className="movie-card__details-item">
-                    <strong className="movie-card__details-name">Released</strong>
-                    <span className="movie-card__details-value">{yearRelease}</span>
-                  </p>
-                </div>
-              </div>}
+            <TabDetails film={film} />
+            }
 
             {activeItem === TABS.reviews &&
-              <div className="movie-card__reviews movie-card__row">
-                <div className="movie-card__reviews-col">
-                  <div className="review">
-                    <blockquote className="review__quote">
-                      <p className="review__text">Discerning travellers and Wes Anderson fans will luxuriate in the glorious Mittel-European kitsch of one of the director&apos;s funniest and most exquisitely designed movies in years.</p>
-
-                      <footer className="review__details">
-                        <cite className="review__author">Kate Muir</cite>
-                        <time className="review__date" dateTime="2016-12-24">December 24, 2016</time>
-                      </footer>
-                    </blockquote>
-
-                    <div className="review__rating">8,9</div>
-                  </div>
-
-                  <div className="review">
-                    <blockquote className="review__quote">
-                      <p className="review__text">Anderson&apos;s films are too precious for some, but for those of us willing to lose ourselves in them, they&apos;re a delight. &quot;The Grand Budapest Hotel&quot; is no different, except that he has added a hint of gravitas to the mix, improving the recipe.</p>
-
-                      <footer className="review__details">
-                        <cite className="review__author">Bill Goodykoontz</cite>
-                        <time className="review__date" dateTime="2015-11-18">November 18, 2015</time>
-                      </footer>
-                    </blockquote>
-
-                    <div className="review__rating">8,0</div>
-                  </div>
-
-                  <div className="review">
-                    <blockquote className="review__quote">
-                      <p className="review__text">I didn&apos;t find it amusing, and while I can appreciate the creativity, it&apos;s an hour and 40 minutes I wish I could take back.</p>
-
-                      <footer className="review__details">
-                        <cite className="review__author">Amanda Greever</cite>
-                        <time className="review__date" dateTime="2015-11-18">November 18, 2015</time>
-                      </footer>
-                    </blockquote>
-
-                    <div className="review__rating">8,0</div>
-                  </div>
-                </div>
-                <div className="movie-card__reviews-col">
-                  <div className="review">
-                    <blockquote className="review__quote">
-                      <p className="review__text">The mannered, madcap proceedings are often delightful, occasionally silly, and here and there, gruesome and/or heartbreaking.</p>
-
-                      <footer className="review__details">
-                        <cite className="review__author">Matthew Lickona</cite>
-                        <time className="review__date" dateTime="2016-12-20">December 20, 2016</time>
-                      </footer>
-                    </blockquote>
-
-                    <div className="review__rating">7,2</div>
-                  </div>
-
-                </div>
-              </div>
+            <TabReviews />
             }
 
           </div>
