+++ src/components/tab-details/tab-details.jsx
@@ -1,0 +1,61 @@
+import React from "react";
+import PropTypes from "prop-types";
+
+const TabDetails = (props) => {
+  const {film} = props;
+  const {producer, actors, runTime, yearRelease, genre} = film;
+
+  return <div className="movie-card__text movie-card__row">
+    <div className="movie-card__text-col">
+      <p className="movie-card__details-item">
+        <strong className="movie-card__details-name">Director</strong>
+        <span className="movie-card__details-value">{producer}</span>
+      </p>
+      <p className="movie-card__details-item">
+        <strong className="movie-card__details-name">Starring</strong>
+        <span className="movie-card__details-value">
+          {actors.map((actor, i, arr)=>{
+            return <React.Fragment
+              key={actor}>
+              {actor}{i === arr.length - 1 ? `` : `,`} <br />
+            </React.Fragment>;
+          })}
+        </span>
+      </p>
+    </div>
+
+    <div className="movie-card__text-col">
+      <p className="movie-card__details-item">
+        <strong className="movie-card__details-name">Run Time</strong>
+        <span className="movie-card__details-value">{runTime}</span>
+      </p>
+      <p className="movie-card__details-item">
+        <strong className="movie-card__details-name">Genre</strong>
+        <span className="movie-card__details-value">{genre}</span>
+      </p>
+      <p className="movie-card__details-item">
+        <strong className="movie-card__details-name">Released</strong>
+        <span className="movie-card__details-value">{yearRelease}</span>
+      </p>
+    </div>
+  </div>;
+};
+
+TabDetails.propTypes = {
+  film: PropTypes.shape({
+    title: PropTypes.string.isRequired,
+    screenshotSrc: PropTypes.string.isRequired,
+    posterSrc: PropTypes.string.isRequired,
+    movieCoverSrc: PropTypes.string.isRequired,
+    genre: PropTypes.string.isRequired,
+    yearRelease: PropTypes.string.isRequired,
+    description: PropTypes.string.isRequired,
+    rating: PropTypes.string.isRequired,
+    numberVotes: PropTypes.string.isRequired,
+    producer: PropTypes.string.isRequired,
+    actors: PropTypes.arrayOf(PropTypes.string).isRequired,
+    runTime: PropTypes.string.isRequired,
+  }).isRequired,
+};
+
+export default TabDetails;
