+++ src/components/tab-reviews/tab-reviews.test.js
@@ -1,0 +1,11 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import TabReviews from "./tab-reviews.jsx";
+
+it(`render TabReviews`, ()=>{
+  const tree = renderer.create(
+      <TabReviews/>
+  ).toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
